Restructure client tool
