+++ client/src/main/java/org/dst/client/cmd/DstCmdStarter.java
@@ -5,29 +5,34 @@ import java.util.Scanner;
 import java.util.function.Function;
 import org.dst.client.DefaultDstClient;
 
-public class Main {
+public class DstCmdStarter {
 
   private static final String defaultAddress = "list://127.0.0.1:8082";
 
-  private Parser parser = new Parser();
+  private static DefaultDstClient client;
 
-  private static ProcessResult processResult = new ProcessResult();
-
-  private HashMap<DstOperationType, Function<DstCommandWithType, ClientResult>>
+  private static HashMap<DstOperationType, Function<DstCommandWithType, ClientResult>>
           commandHandlers = new HashMap<>();
 
   public static void main(String[] args) {
     //TODO(jyx) Check the server is open or not
 
     if (args.length == 0) {
-      processResult.client = new DefaultDstClient(defaultAddress);
+      client = new DefaultDstClient(defaultAddress);
     }
     //TODO(jyx) deal with -h 127.0.0.1 -p 8082
 
-    new Main().loop();
+    //register different operation type handler
+    commandHandlers.put(DstOperationType.STRING, new StringHandler(client));
+    commandHandlers.put(DstOperationType.SET, new SetHandler(client));
+    commandHandlers.put(DstOperationType.LIST, new ListHandler(client));
+    commandHandlers.put(DstOperationType.DICT, new DictHandler(client));
+    commandHandlers.put(DstOperationType.TABLE, new TableHandler(client));
+    new DstCmdStarter().loop();
   }
 
   private void loop() {
+    Parser parser = new Parser();
     Scanner sc = new Scanner(System.in);
     while (true) {
       System.out.print("dst-cli>");
@@ -39,7 +44,6 @@ public class Main {
   }
 
   private ClientResult executeCommand(DstCommandWithType commandWithType) {
-    commandHandlers.put(commandWithType.getCommandType(), processResult);
-    return commandHandlers.get(commandWithType.getCommandType()).apply(commandWithType);
+    return commandHandlers.get(commandWithType.operationType).apply(commandWithType);
   }
 }
